Add big fat ass serialised payload test
In the `ApplicationSubmission` service object, where we will handle the submission to the datastore, we are now using the serializers.

I've added a big, "real" (as in, using real DB records instead of doubles) test to ensure the generated JSON payload conforms to the schema (validated using the `LaaCrimeSchemas::Validator`).

We only need to do this once, and this looked like the perfect place to do it. In the other serializer specs we can continue using doubles or non-persisted records, and we should.