CRIMAP-211 Add JSON adapter and serialisers (#189)
* add private methods in submission service to convert payload

* move document creation logic into adapter

* remove unneeded line from spec

* change datastore setting

* Implement modular JSON serializers with jbuilder

This is a WIP, tests have not been implemented yet, and some things are just a basic scaffolding and a blueprint to a possible approach to modularise and make the application JSON payload easier to work with and also easier to test with unit tests.

It follows a very similar approach to what we did for the Review (Check your answers) page, with sections that can be reorganised and tested in isolation, and definition objects that can be reused (Person, Address...)

* add serializers and some specs

* update serializers with new required obkects

* Few code tweaks and finalise tests

I've made a few code tweaks, for example the definitions now can be passed a collection:

`json.codefendants Definitions::Codefendant.generate(kase.codefendants)`

Also, removed the individual definition specs, as these are implicitly tested via through the serializers that use them. For example offence and codefendant definitions are tested in the `CaseDetails` serializer.

A few other tweaks here and there and finalising the `ApplicationSubmission` service tests (with webmock).

* Finalise applicant attributes

* Fix bundle after merge of main branch

* Add big fat ass serialised payload test

In the `ApplicationSubmission` service object, where we will handle the submission to the datastore, we are now using the serializers.

I've added a big, "real" (as in, using real DB records instead of doubles) test to ensure the generated JSON payload conforms to the schema (validated using the `LaaCrimeSchemas::Validator`).

We only need to do this once, and this looked like the perfect place to do it. In the other serializer specs we can continue using doubles or non-persisted records, and we should.

Co-authored-by: Jesus Laiz <zheileman@users.noreply.github.com>