Better handling of incomplete offences
Offences may be left incomplete (i.e. lacking dates, or even without a name) under certain circumstances like the provider saving a draft in the Charge creation page.

This is usually fine and an application with incomplete offences shouldn't usually get to submission.

But just in case, and because as part of local tests and demos this could happen, improve the handling of these incomplete offences by filtering out in the serializers.