CRIMAP-240 Adapt developer tools to datastore
We are now using the datastore for everything post-submission of an application.

Some operations of the “developer tools“ (in the footer, intended for speed up some actions in development/demo environments) relied on local DB so these need to be tweaked to work again with the datastore.

Now, depending on the current application (in progress, or submitted) the operations will perform the expected action on the correct "database".

Please note marking a submitted application as returned will work, but there is no counter in the returned tab (it was removed couple days ago until counters are reworked, part of another ticket).

However trying to update a returned application (re-hydration) is not yet implemented so will blow up if tried.