CRIMAP-235 Spike SAML authentication
Basic SAML auth against LAA Portal (for now using a mock they've deployed that clearly is lacking some stuff).

Using Omniauth with Warden. Not using Devise just yet as it might not be necessary and also I wanted to focus this spike/code on the actual SAML strategy and IdP metadata, and not so much on the handling of the users (aka providers), if these have to be persisted to DB or not, etc.

Clearly some kinks to iron out, error messages and such...
Also currently not using certificate for the metadata, we probably need that for any environment other than the mock.

NOTE: on localhost "real" SAML auth is disabled. So if you check out the branch, you are using out of the box a mock.

Define `OMNIAUTH_TEST_MODE=false` in your `.env.development.local` (and restart server) to perform real calls to the IdP.