CRIMAP-244 Retry application submission failures
This builds on the back of some previous work we had to handle submission errors, but we were redirecting the user to a generic "unhandled error" page.

Now we have a proper design and content for this error page, and also as per service designer, the `Try again` button will re-try the submission, without taking the user back to the `declaration` page.

The user can keep re-trying as many times as they want, or click `back` link to go to the declaration page, or `Back to all applications` to go to the dashboard.

Also as part of this, a minor copy change to the declaration primary button was done, as per ticket CRIMAP-242.

NOTE: I've introduced a `Steps::NoOpAdvanceStep` concern as well as a `NoOpForm` to be used in controllers that don't persist anything, and are only used to control flow by "advancing".
I will raise a separate PR to use this in a few other places to DRY code.