CRIMAP-236 Reimplement dashboard counters
This adds again the counters for `returned` applications, and also additionally `submitted` applications.

Cleanup a lot of code and made it resilient so the whole dashboard will not blow up if the datastore is not available.

NOTE: I've split the `dashboard_spec` into 2, moving the tests related to the application lists to a separate file, as to avoid as much as possible conflicts with a separate open PR.