CRIMAP-248 Persist signed in provider to DB
This is a pre-requirement to enable further work, as we require having a way to persist some provider details across sessions and browsers.

In doing this, I've moved to use Devise as it simplifies some things (we no longer need to deal directly with warden or a custom strategy) and also enable features that we will need like session timeouts (for now I've set it to 48 hours but can be changed later), locking mechanism, and basic audit (last_sign_in_at etc.)

Functionally however this is the same as was before, it is just a bit of refactor, and persisting to DB the provider.