CRIMAP-40 Confirm office account number (#231)
First part of the provider's office account selection.

Currently this is not yet hooked to the general flow. That will come in a separate PR.

Usual controller, form object, decision tree, etc. New `provider` namespace and decision tree to separate concerns.
Only difference here is that at any point in these steps we will have a `current_crime_application` as these steps come before the creation of the application.