CRIMAP-41 Select office code from a list (#232)
* CRIMAP-41 Select office code from a list

Follow-up to PR #231.

Allows the provider to select the office account number from a list. This will persist the selected code to the provider's setting so we don't have to ask again until next login.

Still not integrated in the wider journey, this will come as a separate PR.

* Tweak devise locales

And remove unnecessary `i18n` initializer.