CRIMAP-40 Confirm office account on sign in
If the provider has more than 1 office account, show the confirmation page (implemented in previous PR #231 and #232) upon signing in.

If it has only 1 account, do not show this page.

The selected account will persist across logins, but the provider will be asked again to confirm, in case they want to change it,