CRIMAP-249 Scope in progress applications by office
Now that we have the office account selection (as per previous PRs), the next step is to start using this office code to scope and filter applications.

In this PR as a first step we are storing the office code in any newly created application, and filtering the dashboard (in progress) applications by office code as well.