Show office list on sign in if no office is selected
This is a refinement on top of previous PR #233.

When a provider signs in for the first time, if they have multiple offices, we don't set a default one, instead we show them a list of offices for them to select.
Subsequents signs in will present them the office confirmation page in case they want to change their selection.

If they only have one office, then we default to that one as before, and do not show them the office selection or confirmation pages.