CRIMAP-224 Collect legal rep details in declaration page
Adapt the existing declaration page to collect the legal representative details, who is the one signing the declaration.

Removed the previous "check box" as it is no longer used.

These legal rep details are also stored as "settings" of the currently signed in provider, so next time they get to a declaration page on a new application, these details get pre-populated.

If any of the details are changed in a declaration page, the settings are also updated, and will have effect on new applications (not existing ones).

NOTE: integration with the datastore will be done in a separate PR.