CRIMAP-232 Add date ranges in offences page
A new requirement is needed to allow adding date ranges, with a start and end date, being the end date optional, to the current offences page.

Figma designs (still some open questions to be addressed in follow-up PRs):
https://www.figma.com/file/thkvnDBbQbHlSqzt1TE3lk/Crime-Apply?node-id=3328%3A27560&t=NqyXwuhArytsMxws-4

Note this PR looks like a lot of changes but it is mostly the removal of the old `date` attribute and replacement with a `date_from` (and `date_to` still not used in many places).

To maintain compatibility as much as possible, existing submitted applications will continue rendering the certificate page. Also, applications will keep submitting a "dates" array with just the `date_from` until the schema is changed in a follow-up PR.

NOTE: locally stored applications (not submitted) will need to be updated to add the `date_from`. I didn't create a migration as don't think is a big problem, but what I did was surface in the Check your answers page if an offence is incomplete, similar to what we already do in the charges summary page.