Submit to datastore new offence ranges (#245)
Final follow up to previous PRs.

This uses the updated schema, to push the new ranges format for offence dates.

However sadly old submitted applications can't be read once this is merged. An SQL migration can be run to update those applications or maybe is time to wipe out the datastore and start fresh.

NOTE: the schemas gem is not yet updated in the datastore. Merge this PR after that is done.