CRIMAP-222 Add missing MAAT ID info to CYA (#246)
Add previous MAAT ID and financial change details (depending on case type selection) to the Check your answers page.

Additionally and on the back of recent PR #244 that made the applicant telephone number optional, update the CYA to show `None` if the telephone is left blank so the provider is aware of this. Same as we do with other optional fields.