CRIMAP-253 Confirm item deletion with flash alert
On the back of some recent accessibility conversations, it was raised the suggestion to show a notification when an item, from a list of multiple items, is deleted, so screen readers can read it aloud.

Made possible by extending the `#update_and_advance` method to support a new `flash` kwarg, so this can be used in other steps to show a flash when advancing to next step.

This applies to the co-defendants page (removing an existing co-defendant) and to the offences page (removing an existing offence date).

Additionally and to be consistent with Offences, the co-defendant "Remove" buttons are also numbered, as required for accessibility reasons.