CRIMAP-239 Preliminary approach to re-hydration (#252)
As part of the "returned application journey", on Apply we want to be able to select a returned application in the dashboard, and click a button "Update application" to edit it and re-submit.

The edition in reality is the creation of a brand new application, keeping all the entered information from the original (or parent) application, including same USN.

The provider is then presented with the `Check your answers` page for this new application, with everything pref-filled so they can change anything they are required to.

Then they can submit the application, which will create a new record in the datastore, with a new UUID, but same USN.

This is the first part of probably a few more PRs to finalise this re-hydration, taking a JSON payload and reconstructing a persisted (in Apply local DB) application in draft mode (`in_progress`), including relationships.

In this PR, a basic re-hydration of the top-level application attributes, and the `applicant` (including addresses if any) if included. Some other code tweaks to facilitate this, along with some renaming/refactoring.