CRIMAP-239 Rehydrate case details
Follow-up to previous PR #252.

In this PR the rehydration of the case details (except IoJ that will be done separate) is implemented.

Most of it has been refactoring previous code that coped with case details JSON from the datastore rendered in the application certificate, to use DB instances instead of Structs so is easier to persist back to the database as we keep the same data modelling in both places.
The code has been simplified too.

Still, WIP so some things might not work as expected. Refer to previous PR for more details on how to test this.