CRIMAP-121 Propagate provider email address
On submission, the currently signed in provider email address gets injected into the application to be propagated to the datastore along with the already existing legal representative attributes.

NOTE: on staging we are currently using a somehow lacking SAML mock, which does not return to us any email address. In the interim until we properly integrate with LAA Portal staging environment, a fallback email will be used, just so the schema validates the payload.

Additionally, and as recently discussed, Apply will stop sending a `status` attribute when submitting applications to the datastore. This attribute is redundant on submissions and the datastore was already setting the status to `submitted` anyways.