CRIMAP-239 Implement returned applications parent_id
The PR is a bit large with a changes in some other files as I took the opportunity to refactor some of the old code we had (from the time when the applications were still not purged in the local DB).

On rehydration of returned applications, persist the UUID of the returned application as `parent_id` in the new, in_progress, application, so by the time it is submitted we can propagate this to the datastore.

This `parent_id` can then be used to know things like:
- it is a resubmission (as first-time submissions will have this attribute to nil)
- what application (submitted, persisted in the datastore) this new resubmission was originated from
- link all resubmissions together for a given USN and perform diffs.