Refactor dashboard helpers and rename sort params (#262)
This follows the refactor done to the datastore endpoint:
https://github.com/ministryofjustice/laa-criminal-applications-datastore/pull/32

It turned out to be more noisy than I would have expected, as the params were scattered in multiple places.

I took the opportunity to refactor the helpers we use in the dashboard pages so these are now all together in a `DashboardHelper` and keep the sorting and filtering more contained.

NOTE: I've noticed we don't have enough integration tests for the filtering, sorting, and pagination of the different dashboard tables. This is something we will need to improve over time.