Update developer tools with new return action
Use the new v2 return endpoint to mark an application as returned from the developer tools.

When viewing a submitted application, go to the footer, developer tools, and there will be a button `Return application`.

Returned applications will show in the returned list in the dashboard.

The reason will be added in another PR.

NOTE: it requires a PR currently open in the datastore to be merged in order for the new endpoint to work.