CRIMAP-264 Show return details in certificate page
NOTE: design and copy is still under discussion with service designers but will be some similar incarnation to what is being implemented here.

When an application is returned, rendering its certificate page will show a notification banner in the top of the page with the return details.
These details come from the datastore.

There is a separate ticket to redesign the application overview details (reference number, date stamp, date submitted...)