CRIMAP-255 Handling of SAML metadata/certificates (#267)
* CRIMAP-255 Handling of SAML metadata/certificates

This is a small iteration on top of what we already had with regards to LAA Portal authentication, to be able to consume metadata files locally stored, in addition to remote metadata endpoints (one, or another, they are mutually exclusive).

This reduces the amount of certificates we need to expose as secrets, given the metadata files are not considered secrets and can be stored in the public repo (the certificate they contain is the public one).

On the other side, we still need to handle with care our own (SP - Service Provider) certificate pair (public/private). Not so much the public, but the private. But for the sake of been together, we store both as secrets.

A certificate has been generated to be used for samlmock and Portal DEV. Samlmock will not use it, and we don't have a deployed DEV env so this can be used for local tests against Portal DEV once we are federated.

NOTE: If we can fetch the metadata endpoints directly in the future, then we can remove the need to consume them locally, or leave it as a fallback alternative. At the moment it is not possible due to network policies but it's being looked after.

* Wrap parser in method to avoid repetition