CRIMAP-267 Better handling of application statuses
This is mostly a refactor and tidy up of some code that no longer makes sense because we only have in progress applications locally, and anything else is remote (datastore).

Previously, we had still a few coupling of the `status` values to locally persisted records, this is no longer the case.

To be extremely explicit, the `enum` in the `CrimeApplication` model now only support `in_progress` values so it forbids trying to persist to local DB a record in any other status (to limit mistakes).

Implement (similar to the others we already had) a `Datastore::GetApplication` service object that encapsulates the call to the API and handling of errors (including not found).

Awaiting confirmation from service designer on what to do with `superseded` applications, but for now, and for the sake of this code, I'm treating them as "not found" which means showing a "nice" error page to the user (this is an existing error page).

Finally, fixed a bug in the `ValueObject` superclass where inquiry methods were polluting other value object subclasses namespace.