CRIMAP-255 Use cert-manager for Portal certificates
Follow-up to PR #267.

Instead of having to generate manually (`openssl genrsa...`) the service provider certificates to use with Portal, and then base64 encode them and store in a secrets file, we can use k8s cert-manager to issue these self-signed certs and output directly to secrets.

It has the advantage we don't have to remember the openssl commands, nor git-crypt the file anymore as it does not contain any secrets, and also they can be regenerated or renewed automatically.

If these ever change (or expire) we will still have to provide to PCUAM team our SP metadata file containing the new (public) certificate.