Prepare Apply to use datastore JWT auth
Merging this will not enable JWT or change any functionality on Apply, as it is not enabled yet (to enable, change `config.auth_type = :jwt`).

Once enabled, the datastore API client will make use of a little middleware introduced in the `simple-jwt-auth` gem to generate and send a JWT bearer token.