Set a more sensible default for ENV variable (#279)
With openssl >= 3.0.0 you need to have something in those secrets. Nil or empty string will produce an error.

With openssl < 3.0.0 a nil value will blow up but funnily enough not an empty string.

To avoid confusion and edge cases with the openssl version installed in each system, we set a default dummy value, so developers don't need to set their own in their `.local` files.