CRIMAP-254 Avoid error message repetition
As a preparation for an accessibility audit, we must try to avoid repetition of error messages (in summary) when in a page there are more than "1 thing of the same".

For instance, in the interests of justice page, when more than one justification text area was left blank, the error summary was repeating "Enter more detail" multiple times. This is no good.

Tweaked these and other minor changes in Offences page (hint texts, labels, etc.) after suggestions from content designer.

Some more copy changes might be implemented on follow-up PRs.