CRIMAP-296 IoJ secondary input field copy tweaks
Related to accessibility issue "Secondary input field not linked to primary input".

We are using visually-hidden labels for each of those text areas, so I believe a screen reader should have additional context when using each of the text areas.

For example these visually hidden labels read like this:

- Unable to represent themselves justification
- Expert cross-examination justification

Etc.

I've updated to copy in some of them to be in line with the same copy we use in the Check your answers page.