CRIMAP-172 Use non-cluster specific domain
As per documentation:

https://user-guide.cloud-platform.service.justice.gov.uk/documentation/other-topics/migrating-from-live-1-domain-name.html

This is part of a broader ticket to use a custom domain for staging, however, for now this makes the service more resilient in case of cluster changes, until we have a production namespace and we set the proper final (and public) domains.

Old `*.apps.live` domain redirects to new domain to avoid breaking bookmarks etc.