CRIMAP-298 Ensure DWP mock works in all cases
There was some scenarios were the DWP mock values wouldn't work, if the day or the month was entered with a different format to the expected by the mock.

Tweaked the code so it works no matter the format, so it works for 1-digit or 2-digits days and months (i.e. 1/1/1980 or 01/01/1980).