CRIMAP-38 Replace samlmock with portal staging (#292)
Swap SAMLmock with Portal Staging.
So our staging talks to Portal staging, as it should be.

Keeping the DEV and samlmock metadata files around, because on localhost envs is useful to be able to use SAMLmock, although ideally this will some day be replaced with DEV (but it is not working so far).

There will be some follow-up PRs for some tickets that are unblocked by this PR.

NOTE 1: there is still a chance this will not work on staging. I've tested locally SP-initiated auth and it works, but can't test IdP-initiated auth until this is deployed and our service is using the staging certificates.

NOTE 2: if this works, new credentials are required to sign in to Apply. These credentials will be shared with the team. Also, the office codes of the new user(s) will change so old applications will not be accessible from Apply anymore. We can create a migration task for the datastore to re-assign office codes.