CRIMAP-246 Provider's office account enrolment check
Implement an initial enrolment/allowlist based on office account codes.

The error page where the user is redirected is still TBD and copy and design will be provided once that has been agreed (there is a separate ticket).

Upon some comments in previous PR #294 I cleanup some of the previous gatekeeper code, and did some renaming too.

Note it may be we also want to allow by email address, but that has not been decided yet. If required, it is almost ready to implement, only caveat we would not want emails committed to the repo, so they might require hashing.