CRIMAP-300 Tweaks for Portal SLO
This is a follow-up to PR #293 for the SLO investigations/spike.

Trying out `post` binding instead of `redirect` binding, as with redirect we get blocked with a 403 by cloudfront.

Also, do not blow up at the rack level if the endpoint can't process the SLO request (for example if it is called without a SAML payload), and show a nice page with a flash error instead (by bubbling up a re-raised `StandardError`).