CRIMAP-305 Fix govuk branded assets
Some assets coming from `govuk-frontend` were not being correctly referenced and part of the assets pipeline.

These mainly includes favicons of different sizes.
Removed static favicon files in `public` directory as we don't want these to be used, but instead the govuk frontend ones.

Modernised the `govuk_template` and the references to these paths.

Additionally, set logging to stdout on production (this is separate to lograge that logs the requests and was already configured to use stdout), and add some additional logging in key areas.