CRIMAP-294 Navigation landmarks healthcheck (#304)
* CRIMAP-294 Navigation landmarks healthcheck

Part of the accessibility improvements ahead of the audit.

Ensure all navigation landmarks have unique labels.

We use some navigation components (`<nav>`). Some of these required tweaking and locales.
This will assist screen reader users to quickly navigate to the content they wish to access.

* Remove unneeded fallback image

This image line is not required as it was only useful as fallback for IE 8, which we don't support.