Update client details task completed rules
Originally we had the `telephone_number` as a mandatory field, but some time ago we decided to make it optional, thus we don't have to check for presence in order to consider the task "completed", as `telephone_number` can be left blank.

Only `correspondence_address_type` is mandatory now.

Made the rules for the correspondence address more stringent by checking we have indeed the address record filled when `home address` or `other address` are selected.