CRIMAP-307 Mitigate header menu flickering
This was driving me crazy when testing the site on small screens / mobile.
It is very noticeable on Safari, less so on Chrome. The recordings are from Safari.

There is an annoying thing happening with the header menu when a user is signed in to Apply.

On small screens, like mobile devices, each time a page is loaded, the header menu briefly shows (a few milliseconds) before being hidden. This is a bit annoying.

The header menu should be hidden on small screens unless the user clicks the “Menu“ button, at that point the menu is revealed, until they click the “Menu” button again, and it is hidden.

After playing with the order of the javascript loading, `preload` and other things, I've not been able to pinpoint exactly why this happen and how to solve it, thus this mitigation which shouldn't have any adverse effect as far as I can tell.

With this fix the Menu link still flicks a bit on Safari, but at least the page content doesn't jump around.

Note: desktop is not affected. This only affects small screens where the menu is collapsed to save space.
Also, when javascript is disabled, menu shows expanded by default on page load, same as before.