Refactor locales and simplify partial
Reorganised the code in the partial to remove conditionals and let the locales handle the different combinations.

This makes it more flexible if new reasons are added with different explanations, call to action, etc.

There is now a bit of repetition on the heading locales, "LAA have returned this application because [...]" but IMO makes it clearer, and probably easier to translate/tweak in the future.