CRIMAP-331 Refactor IoJ and means passporters
This PR is a follow-up to some edge cases detected on previous PR #317.

Previously we had only one IoJ passporter. Now we have another passporter named `MeansPassporter` to better encapsulate the logic and meaning of means passporting.

Means passporting can be done on age (under 18s) or if over 18, then based on DWP check, that can result on success or fail.

Refactor some of the previous code, moving passporter classes into its own namespace `Passporting`.
Removed the `AgeCalculator` and converted to a concern `PersonWithDob`.

Fixed the task list, that was still relying on NINO being a mandatory field (it is not anymore, as under 18s are exempt of DWP check and thus we skip the NINO page).

Minor DRY tweaks along the way.

NOTE: this is the first part of this work. In a separate PR, this new means passporting attribute will be exposed in the JSON for submission, so review can use the resulting values, and also Apply can read back these.