CRIMAP-337 Deal with returned passported applications (#323)
One edge case not covered in previous PR was the fact, a passported application (mostly on the basis of under 18, but could also be on the basis of the DWP check) that is returned back to the provider, should keep the original passporting and not try to re-run the checks or take the provider through a different journey to the original one.

In practice this means:

- if original application was for an under 18 client, then the returned application will not ask for NINO nor run the DWP check, and remain passported upon submission.
- same for over 18 with a successful DWP check.

This PR introduce some refactor to make this easier, with some syntax-sugar methods in a concern `Passportable`.

Also, introduced (although enabled on staging) a feature flag to enable/disable the means passport on under 18 (i.e. ask or not for NINO + DWP check). Just in case we need to turn this off.

Additional edge cases need to be studied and covered with code if required, as follow-up PRs.
Also some changes might be required to the IoJ Passporter once we introduced the offence passporting.