CRIMAP-326 Prepare for deployment to production (#329)
Added manifest files for the new production namespace.

DRY github workflow to allow for reusable deploy action with parameterised inputs.

NOTE: using for now DWP staging endpoint and creds. This will be changed once we have prod ones.

Also: using http basic auth (different to the one on staging, will share with the team). LAA Portal will not work as we are not onboarded, but all necessary for the onboarding will be created with this PR.