Add staging subdomain to ingress
This is a subdomain of our production domain.

Internal cloud platform URLs are still available and functional, but this is now an easier to remember domain.
The cloud URLs will redirect to their corresponding "nice" domains.

Some functionality in Portal will still redirect to old cloud URL until metadata is updated on their side.