Tweak devise session timeout depending on env
We had a hardcoded, long session timeout. This was fine for staging as we don't want to have to continuously authenticate with Portal (specially because it also needs the VPN).

However in production we want this to be a much shorter period of time, to increase security.

It is configurable, so in localhost we can still have a longer or shorter period that can help with development by setting the `SESSION_TIMEOUT_MINUTES` variable in the `.env.development.local` file.

Reduced staging from 48 hours to 24 hours timeout.
Configured production to 1 hour timeout.
These are sensible defaults, can be tuned later on.