CRIMAP-346 Under 18 IoJ on returned applications
This covers an edge case where the applicant is aged under 18 and thus is passported on IoJ on the first submission and then the application is returned by the caseworker when applicant is already over 18, or the applicant turns 18 before the application is resubmitted again.

In summary, if the application is a resubmission, we don't run again the age check, instead we use whatever result the initial submission had.