CRIMAP-355 Implement session lifespan (#337)
This is in addition to the existing idle session timeout mechanism.
Implemented via a new custom Devise module `:reauthable`.

The session lifespan will sign out a user (provider), and force them to re-authenticate with Portal, after X minutes have passed from their last sign in, regardless of their activity.

It is configurable, so in localhost we can still have a longer or shorter period that can help with development by setting the `REAUTHENTICATE_AFTER_MINUTES` variable in the `.env.development.local` file.

Staging set to 48 hours. Production set to 24 hours.
These are sensible defaults, can be tuned later on.