CRIMAP-146 Prepare for offence IoJ passporting
Still awaiting final CSV file that will mark the offences that are "passportable". This PR just assumes a new column, named `ioj_passport` and for now it is `nil` for all offences, thus none is passportable.

The logic however is implemented in this PR and also a feature flag to globally (instead of at the offence level) can enable/disable this IoJ passporting.

I've not wanted to make changes to our current CSV file because it will have to change soon again, probably the columns too, as we are splitting multi-class offences.