CRIMAP-21 Add basic prometheus rules
This is a first bunch of basic prometheus rules that will alert us in the alerts channel whenever they fire.

There is scope to add some more or tweak them so for now I'm adding them here in the code repo, but once we are comfortable with them, we should move this file from the code repo to the infra repo (`cloud-platform-environments`).