CRIMAP-9 Add some custom GA events
This will enhance the basic analytics, with some custom events triggered for certain actions throughout the application steps.

Radio options, checkboxes and dates (just the year part) will be submitted as events on form submits.

Also, individual links can be annotated with the class `ga-pageLink` and a couple data attributes to also send an event.
This is useful mostly for "action" links, for instance "change", "remove"... links.
I've added a few examples but we can annotate more links if they seem useful.